backend/requirements.txt
Flask==3.0.3
pyserial==3.5
requests==2.32.3
